backend/pyproject.toml
[tool.poetry]
name = "sorttea-backend"
version = "0.1.0"
description = "Backend service for managing and verifying Instagram giveaway entries"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
celery = "^5.3.6"
redis = "^5.0.1"
gunicorn = "^21.2.0"
django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-django = "^4.7.0"
black = "^23.12.0"
isort = "^5.13.2"
flake8 = "^6.1.0"
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 